refactor: 유효성 검사 상수화 파일 위치 변경
